separated socketIO code
